Fix path with accents for tvpaint
