+++ openpype/hosts/tvpaint/api/pipeline.py
@@ -2,17 +2,13 @@ import os
 import json
 import tempfile
 import logging
-import sys
 import requests
 import pyblish.api
-import ctypes
-from ctypes import wintypes
-from pathlib import Path
 from openpype.client import get_asset_by_name
 from openpype.host import HostBase, IWorkfileHost, ILoadHost, IPublishHost
 from openpype.hosts.tvpaint import TVPAINT_ROOT_DIR
 from openpype.settings import get_current_project_settings
-from openpype.lib import register_event_callback
+from openpype.lib import register_event_callback, optimize_path_compatibility
 from openpype.pipeline import (
     legacy_io,
     register_loader_plugin_path,
@@ -146,7 +142,7 @@ class TVPaintHost(HostBase, IWorkfileHost, ILoadHost, IPublishHost):
 
     # --- Workfile ---
     def open_workfile(self, filepath):
-        filepath = self.optimize_path_compatibility(filepath).replace("\\", "/")
+        filepath = optimize_path_compatibility(filepath).replace("\\", "/")
         george_script = "tv_LoadProject '\"'\"{}\"'\"'".format(
             filepath
         )
@@ -158,39 +154,9 @@ class TVPaintHost(HostBase, IWorkfileHost, ILoadHost, IPublishHost):
         context = get_global_context()
         save_current_workfile_context(context)
         # Execute george script to save workfile.
-        filepath = self.optimize_path_compatibility(filepath).replace("\\", "/")
+        filepath = optimize_path_compatibility(filepath).replace("\\", "/")
         george_script = u"tv_SaveProject {}".format(filepath)
         return execute_george(george_script)
-
-    @staticmethod
-    def optimize_path_compatibility(filepath):
-        workfile_path = Path(filepath)
-        workfile_path.parent.mkdir(parents=True, exist_ok=True)
-
-        if 'win' not in sys.platform:
-            log.info("improve_compatibility: nothing done, only applicable for Windows.")
-            return filepath
-
-        # Windows-specific logic to convert the filepath to its short path form.
-        _GetShortPathNameW = ctypes.windll.kernel32.GetShortPathNameW
-        _GetShortPathNameW.argtypes = [wintypes.LPCWSTR, wintypes.LPWSTR, wintypes.DWORD]
-        _GetShortPathNameW.restype = wintypes.DWORD
-
-        workfile_parent = str(workfile_path.parent)
-        output_buf_size = len(workfile_parent)
-        # Iteratively try to get the short path name, increasing buffer size if needed.
-        while True:
-            output_buf = ctypes.create_unicode_buffer(output_buf_size)
-            needed_size = _GetShortPathNameW(workfile_parent, output_buf, output_buf_size)
-            if needed_size == 0:
-                raise ctypes.WinError()
-            if output_buf_size >= needed_size:
-                return os.path.join(output_buf.value, workfile_path.name)  # Return the short path version.
-            else:
-                output_buf_size = needed_size  # Adjust the buffer size if needed.
-
-        # Fallback return to the original path if for some reason Windows-specific logic fails.
-        return str(workfile_path)
 
     def work_root(self, session):
         return session["AVALON_WORKDIR"]
@@ -342,7 +308,7 @@ def get_workfile_metadata_string_for_keys(metadata_keys):
         mode="w", prefix="a_tvp_", suffix=".txt", delete=False
     )
     output_file.close()
-    output_filepath = output_file.name.replace("\\", "/")
+    output_filepath = optimize_path_compatibility(output_file.name).replace("\\", "/")
 
     george_script_parts = []
     george_script_parts.append(
