+++ openpype/hosts/tvpaint/plugins/load/load_image.py
@@ -1,7 +1,7 @@
 from openpype.lib.attribute_definitions import BoolDef
 from openpype.hosts.tvpaint.api import plugin
 from openpype.hosts.tvpaint.api.lib import execute_george_through_file
-
+from openpype.lib import optimize_path_compatibility
 
 class ImportImage(plugin.Loader):
     """Load image or image sequence to TVPaint as new layer."""
@@ -77,7 +77,7 @@ class ImportImage(plugin.Loader):
         )
         # Fill import script with filename and layer name
         # - filename mus not contain backwards slashes
-        path = self.filepath_from_context(context).replace("\\", "/")
+        path = optimize_path_compatibility(self.filepath_from_context(context)).replace("\\", "/")
         george_script = self.import_script.format(
             path,
             layer_name,
