+++ openpype/hosts/tvpaint/plugins/load/load_reference_image.py
@@ -1,6 +1,7 @@
 import collections
 
 from openpype.lib.attribute_definitions import BoolDef
+from openpype.lib import optimize_path_compatibility
 from openpype.pipeline import (
     get_representation_context,
     register_host,
@@ -86,7 +87,7 @@ class LoadImage(plugin.Loader):
         subset_name = context["subset"]["name"]
         layer_name = self.get_unique_layer_name(asset_name, subset_name)
 
-        path = self.filepath_from_context(context)
+        path = optimize_path_compatibility(self.filepath_from_context(context))
 
         # Fill import script with filename and layer name
         # - filename mus not contain backwards slashes
