+++ openpype/hosts/tvpaint/worker/worker_job.py
@@ -9,7 +9,7 @@ from abc import ABCMeta, abstractmethod, abstractproperty
 
 import six
 
-from openpype.lib import Logger
+from openpype.lib import Logger, optimize_path_compatibility
 from openpype.modules import ModulesManager
 
 
@@ -213,13 +213,13 @@ class ExecuteGeorgeScript(BaseCommand):
             )
             output_file.close()
             format_key = "{" + key + "}"
-            output_path = output_file.name.replace("\\", "/")
+            output_path = optimize_path_compatibility(output_file.name).replace("\\", "/")
             script = script.replace(format_key, output_path)
             filepath_by_key[key] = output_path
 
         # Replace job queue root in script
         if self._root_dir_key:
-            job_queue_root = self.job_queue_root()
+            job_queue_root = optimize_path_compatibility(self.job_queue_root())
             format_key = "{" + self._root_dir_key + "}"
             script = script.replace(
                 format_key, job_queue_root.replace("\\", "/")
@@ -384,7 +384,7 @@ class SenderTVPaintCommands(TVPaintCommands):
 
         It is expected that worker will add it's root before passed workfile.
         """
-        new_workfile = workfile.replace("\\", "/")
+        new_workfile = optimize_path_compatibility(workfile).replace("\\", "/")
         job_queue_root = self.job_queue_root().replace("\\", "/")
         if job_queue_root not in new_workfile:
             raise ValueError((
@@ -475,8 +475,8 @@ class ProcessTVPaintCommands(TVPaintCommands):
 
     def _prepare_workfile(self, workfile):
         """Preprend job queue root before passed workfile."""
-        workfile = workfile.replace("\\", "/")
-        job_queue_root = self.job_queue_root().replace("\\", "/")
+        workfile = optimize_path_compatibility(workfile).replace("\\", "/")
+        job_queue_root = optimize_path_compatibility(self.job_queue_root()).replace("\\", "/")
         new_workfile = "/".join([job_queue_root, workfile])
         while "//" in new_workfile:
             new_workfile = new_workfile.replace("//", "/")
@@ -507,7 +507,7 @@ class ProcessTVPaintCommands(TVPaintCommands):
         )
         temporary_file.write(george_script)
         temporary_file.close()
-        temp_file_path = temporary_file.name.replace("\\", "/")
+        temp_file_path = optimize_path_compatibility(temporary_file.name).replace("\\", "/")
         self.execute_george("tv_runscript {}".format(temp_file_path))
         os.remove(temp_file_path)
