+++ openpype/lib/__init__.py
@@ -174,6 +174,10 @@ from .connections import (
     requests_post
 )
 
+from .path_tools import (
+    optimize_path_compatibility
+)
+
 terminal = Terminal
 
 __all__ = [
@@ -272,6 +276,8 @@ __all__ = [
     "version_up",
     "get_version_from_path",
     "get_last_version_from_path",
+
+    "optimize_path_compatibility",
 
     "merge_dict",
     "TemplateMissingKey",
